Created webView functionality and new label for custom cell
Created a webView instance that gets called if the user clicks a button
on the detailView page.  This calls the manufacturer specific site in a
pop-over ‘view’ and can be dismissed by clicking the button again.
Also added a country of origin label to the custom cell to satisfy
project requirements.