Added new custom cells for TableView and formatting
Created custom cells that look like Monster U id cards for the
tableview.  Formatted it so the background is transparent and adds
spacing as well.  Also added in custom icons for the nav bar controller
on the bottom.  Created an addtional entry in the tableView to hold an
“Additonal Characters” entry which displays custom bios from an article
found on Full Sail Library site to satisfy project requirements.