HUGE overhaul...new CreateMonster tab and NSUserDefaults functionality
Realized that I misinterpreted the directions and needed to create some
sort of form UI to save defaults, which the “Cast” tab was originally
supposed to be.  So I created another tab for a user to create their
own monster which utilized the Custom Object class to save
NSUserDefaults, and the user can then open the createdMontster view
which loads in the monster that is saved via the same Custom Class
Object.  Whew.