Code added to remove locations from custom class.  Also made annotations for detailView
Created NSMutableArrays within the custom BusinessClass to hold the
names, latitudes, and longitudes of each location so that when one is
deleted within the tableView, it is also removed from the arrays within
the custom class.  This is needed for the full map view tab later.
Also implemented code to pass these parameters to the DetailView
controller for the creation of a pin annotation on the mapView.