created parser functionality and made newsObject
Finalized the parser functionality to recognize the names of the
returned articles from the rest of the crap that is returned (and
unneeded).  Also created a custom NewsObject to contain all data and
functionality so that it can handle the retrieval, storage, and
manipulation of its own data to keep things more clean.  Also passed
the entire returned xml data as a string to the xml View to display, as
per the project requirements